requirements.txt
python-dotenv
pycoingecko
google-generativeai